requirements.txt
streamlit==1.41.0
pandas==2.2.2
scikit-learn==1.4.0
matplotlib==3.8.4
seaborn==0.12.1
xgboost==2.0.3
streamlit-reveal-slides==0.1.0

